Add models for consent groups
Add models to track data use permission and modifiers from the
Data Use Ontology (GA4GH). Add a model for study consent groups.
Note that we will store the full consent code as well as the
data use permissions and modifiers separately, because we'll need
to use the full code to assign access for a dbGaP workspace. The
dbGaP codes don't necessarily match the DUOs codes.