Extend workspace_import template to allow autocomplete
The dbGaPWorkspaceDataForm uses an autocomplete widget, but none of
the autocomplete javascript was loaded in workspace_import. (Note
that it *is* loaded in workspace_form/WorkspaceCreate view because
the WorkspaceForm from ACM uses autocomplete, but the
WorkspaceImportForm does not.) Therefore we need to extend the
workspace_import template to load it for the workspace_data_formset.