Extend the workspace_detail.html template for dbGaPWorkspaces
Ideally this would be a separate template in the dbgap app, but
right now you can't specify a template in the ACM workspace adapter.
Once that's implemented, we can remove this hacky template and
extend the workspace_detail.html template for the dbGaP workspace
types specifically in the dbgap app. Note that I've created an
issue for this (#30).