Add a dbGaPStudy model and a foreign key in dbGaPWorkspace
We don't want to duplicate the study-phs link in every dbGaPWorkspace
because it is fixed. Instead, we should have the dbGaPStudy model
with phs and an fk to study, and then dbGaPWorkspace will have an
fk to dbGaPStudy. Note: we had discussed having another intermediate
dbGaPStudyVersion model and decided that it wasn't worth it. We can
always have a data migration if we find out that we do want it as
a separate table. Modify tests, the form, and the table associated
with dbGaPWorkspace.