Add a form to validate DAR info from dbGaP
This json comes from a dbGaP webpage where a user can get the list
of approved DARs for a given set of dbGaP application project ids.
The user will likely need to navigate to this page, copy the json
text, and paste it into the Django form on this view. (It would be
nice if Django could automatically pull the json with a GET request,
but that can only happen if Django can log into the eRA commons/dbGaP.
It is unlikley that this can happen.) Add a json schema so that
we can validate the text that is cut and pasted. Note that this
schem and form will only allow *one* project ID, while the dbGaP
website allows multiple. We may want to change this in the future
to handle multiple IDs, but for the current purposes of this form
(getting DARs for a specific dbGaP application project ID), it
should only allow one.