Store the current DAR status in dbGaPDataAccessRequest
We'll likely want to know which DARs an application contained
whether or not they were approved. This will also help later when
we need to update a dbGaP application's current DAR statuses (and
therefore workspace sharing). Add this field, a migration to create
it, and modify tests as needed to check this field.