Add simple history tracking to dbgap models
Add a history field to each model and a migration to create that
table/field in the database. Add django-simple-history to installed
packages and requirements file. Modify the admin interface so the
admin classes inherit from the SimpleHistoryAdmin, which allows us
to see the history in from the admin site.