Add a ManagedGroup when creating a dbGaP application
This group is intended to provide access to workspaces that a dbGaP
application is approved to use. Eventually, we intend to build an
interface that lets the PI manage membership. For now, the CC will
likely manage membership in this group based on what the PI indicates.
To implement this group, add the anvil_group field to dbGaPApplication
and an associated migration. Modify dbGaPApplicationCreate to create
a new group in the app, make an API call to create it on AnVIL, save it,
and link it to the newly created dbGaP application. Add a link to the
group detail page on the dbGaP application detail template.