Store original version and participant set for DARs
Store the original version and participant set for dbGaPDataAccessRequests
with the same dbgap_dar_id as a dbGaPDataAccessRequest that already
exists. This is because dbGaP does not tell us (or store?) the
original version/participant set of study that a DAR grants access to.
Instead, we need to look up the current version the first time we
create a DAR. The next time we update that DAR (and create a new
dbGaPDataAccessRequest object), we need to get the version of the
phs from the previous DAR, not look up the now-current version.

Note that this makes the database denormalized, but since we're
relying on external data from dbGaP it seems safer to store it
again instead of assume what should be constant. Instead, we can
check the things that we think should be constant for a given
DAR_ID and raise an exception if they are not.