Rename version and participant set to include "original"
For DARs, we are tracking the original version and participant set
that that DAR grants access to, and it is not coming from the JSON.
Rename dbGaPDataAccessRequest fields dbgap_version to original_version
and dbgap_participant_set to original_participant_set. Note that
the dbGaPWorkspace model still uses dbgap_ for these variables.