Update has_access for changes in ACM
django-anvil-consortium-manager implemented a has_access method
that checks if a Group has access to a Workspace and vice versa.
Use this method in the dbGaPDataAccessRequest has_access method.
Also rename it to "has_access" because it is checking things in
the app, not on AnVIL (although they should be in sync).