Modify has_access to only look at authorization domains
Our plan for PRIMED is to share the workspace with PRIMED_ALL and to
add the dbGaP groups to the auth domain for the workspaces that they
should have access to. The "has_access" method for DARs should
only check if the group is in the auth domain of the matched workspace,
not if the workspace is shared with the group.