Show tables on the dbGaP data access audit page
Add an audit results table class to the audit source code, and methods
for dbGaPDataAccessSnapshotAudit to create tables for the various
types of results that it stores. Display the tables on the audit
page in dropdowns, similar to the anvil auditing functionality. Note
that we will want to show links to grant/remove access where
appropriate, but that is not implemented yet.