Only show alert on snapshots that are not the most recent
The dbGaPDataAccessSnapshotDetail page showed an alert that the
snapshot was not the most recent snapshot available, even if it
was the most recent. Fix it by adding the if is_most_recent check
back into the template. Add tests to both the detail view and the
audit view to check if appropriate text is or is not in the response.