Filter to approved DARs when setting original_version and participant set
Previously, the code looked up version and participant set from
any previous DAR with the same dbgap_dar_id. In the situation where
version and participant set got set for a "new" DAR and the study
updated to a new version before the DAR got approved, we would pull
the version and participant set from the "new" DAR. It would therefore
be incorrect. Modify the code to filter to DARs that have the same
dbgap_dar_id *and* are approved, and if there are none, then query
dbGaP for the current version.