Add "home" to the LOGIN_REQUIRED_IGNORE_VIEW_NAMES setting
login_required.middleware.LoginRequiredMiddleware sets all views to
be LoginRequired except for those specified in settings (the
LOGIN_REQUIRED_IGNORE_VIEW_NAMES array). When not logged in, the "home"
url redirects to `accounts/login/?next=/`, which then redirects
back to `/` using the next parameter after logging in. We want the
login to go directly to the user's profile page, and having the
next parameter in the url skips the LOGIN_REDIRECT_URL setting
(which is set to go to the user's profile page). Try fixing this by
not requiring login to go to the homepage.

There may be a better solution, though, eg a link to the login
page if the user is not logged in?