Modify JSON schema for updating multiple applications in the future
Rename the JSON schema that are being used to have JSON schema for
a given project and JSON schema for what dbGaP returns, ignoring the
max number of items in the array. Move the check for only one
item when creating a dbGaPDataAccessSnapshot to the form clean
method instead of hardcoding it into the JSON. Modify other methods
as necessary to account for the renaming.