Add a link to the latest snapshot in dbGaPApplication list
This was a little frustrating. The way I had the column set up, if
I tried to pass the linkify kwarg, the str formatting would change
from the pretty date-time formatting to the object string itself.
If render_last_update was set, then it did not even show a link. So,
subclass the ManyToManyColumn to create a ManyToManyDateTimeColumn
object. In that object, define the transform method to render the
DateTime column template. Then use that column in the dbGaPApplication
table, pass the filter keyword to subset to just the most recent,
and pass linkify_item=True. This seems to have worked.