Ensure media is loaded and in the proper order for select2
The select2-bootstrap-5-theme css needs to be loaded after select2
css. The widgets that come with django-autocomplete-light do not
include the select2-bootstrap-5-theme, so that needs to be attached
to the form. Add a form mixin that provides this media and use it
in all forms using django-autocomplete-light select2 widgets. As
part of this, I discovered that the workspace edit templates also
need to be loading the workspace_data_formset media as well as the
form media. Override the javascript template blocks for those templates
to ensure the workspace_data_formset media is loaded.