Change auditing to be on a dbGaPApplication, not a snapshot
When adding auditing for a specific dbGaPWorkspace (vs for a specific
dbGaPApplication), it will be a better parallele if the dbGaPApplication
auditing works on the application instead of on a snapshot. Modify
the auditing class to pull the latest snapshot for a given
dbGaPApplication and audit that access. Add a new view to show
audit results for a dbGaPApplication and remove the view that showed
results for a given snapshot. Modify the dbGaPApplication detail
template to show a button to audit the application (using the most
recent snapshot), and remove the "audit snapshot" button from the
dbGaPDataAccessSnapshotDetail template. Also change the table of
dbGaPDataAccessSnapshots to be within an accordian-style dropdown
since it will eventually get pretty long.