Allow access auditing of dbGaPWorkspaces by application
Add an audit method that operates in a single workspace and loops
over all dbGaPApplications to check access (using current snapshot).
Add a view that runs the audit for a specific workspace and renders
a results table. Add an associated url and template. Separate the
results table rendering code into its own template and include it in
both the dbgapworkspace_audit and dbgapapplication_audit templates.