Add a form and view to create a new MemberAgreement
This view creates both the SignedAgreement and the associated
MemberAgreement. Add a form for SignedAgreement as well as for
MemberAgreement. Some other changes were necessary: add a validator
for the SignedAgreement cc_id field. Modify the str method for
AgreementVersion to include a v, which is a little more readable.
Modify the clean method for AgreementType to check that it has an
associated signed_agreement.