Use a single setting to determine AnVIL group names
Instead of having separate settings for the CDSA groups and dbGaP
groups on on AnVIL, use one setting and add the "DBGAP" and "CDSA"
bits to the prefix in the respective views. This will be easier to
maintain in the long run since we don't have to keep adding new
settings for prod/dev/test, especially if there are more data access
mechanisms in the future.