Use new mixin to create MemberAgreements
Add a new Mixin based on the MemberAgreementCreate view to create
a new signed agreement and the corresponding type. Right now it is
only used for the MemberAgreementCreate view, but it will be used
for the other create views as well (but after some changes, which
will be required to make it more flexible).