Modify new Mixin for use with other agreement type create views
Use the new AgreementTypeCreateMixin in the views to create other
agreement types. Modify the mixin such that it works with both these
views. In particular, add additional methods called by form_valid to
create the agreement object, to create the agreement type object
(which needs an anvil group in addition to the info in the form for
data affiliate agreements), and to create resources on AnVIL (again,
since the data affiliate agreement needs to create a the additional
upload group).