Change auditing to work on SignedAgreements only
There isn't the same level of auditing an agreement and a workspace
together. Instead we'll want to audit agreements (to do some basic
checks and verify that they are or are not in the SAG group on AnVIL)
and separately audit workspaces to verify that there is a valid
SignedAgreement for that study's data.