Restructure cdsa audit source code
Create an audit subdirectory. Move the current audit.py file to a
specific file for signed_agreements in that subdirectory. Rename some
classes in that source file to have names that are not specific to
SignedAgreements, because the class names are getting long and can
be separated by their source file. (Note that we will add auditing
for CDSA workspaces next, so separating the code into different
files is useful.)