Add a view to display audit results for SignedAgreements
This view is very similar to the view for auditing a dbGaP workspace
or application. Because some of the template is the same (the
audit_results_tables part), move the __audit_tables.html template
to the main templates directory, then use this template for
displaying any audit results (dbGaPApplication, dbGaPWorkspace, and
SignedAgreements).