Show agreement group in SignedAgreement tables
Add a new property, agreement_group, to the SignedAgreement model.
This pulls out the particular agreement from the associated
AgreementType model. Also add a new method, get_agreement_type(),
which returns the specific AgreementType for a given SignedAgreement.
Then add a column for the agreement_group in the overall
SignedAgreementTable table and the SignedAgreementAccessAudit table.