Do not require login for CDSA records views
These views should be open to all, not only logged in users. Add the
url names to the LOGIN_REQUIRED_IGNORE_VIEW_NAMES list in the settings
file. Remove the LoginRequiredMixin mixin from the records views.
Modify the tests such that the CDSA records tests test that an
unauthenticated user still gets a 200 response code.