Show multiple matching workspaces in the DAR table
Instead of assuming that there is only one workspace that can match
a given DAR, use a column that expects a queryset instead of a single
workspace. Combine the has_access and workspace columns into a single
column that shows both the workspace name and an indicator of whether
the DAR has access.