Replace existing major_version field with the new foreign key
Remove the previous major_version field and rename the temporary
major_version_fk field to major_version. Note that a lot of the code
will be broken since we didn't udpate any of it yet for these changes.
Rename the constraint, since mysql/mariadb will likely complain if
we don't.