Update AgreementMajorVersionDetail to be a DetailView
Instead of having AgreementMajorVersion be a template view that
operates on a non-existent model class, use a detail view that
pulls and instance of the AgreementMajorVersion model. Add a
get_absolute_url method to AgreementMajorVersion, and remove the
get_major_version_absolute_url method from AgreementVersion since
it is no longer necessary. Update tables and templates as necessary
for the removal of this method.