Add view to update a SignedAgreement status
Add a view that updates the status of a SignedAgreement. Add three
different urls that use this view, one for each agreement type. Note
that the the different urls need to pass the correct agreement_type
as a kwarg to the view. I decided to do this because then the urls
are more consistent with the detail views, eg:
cdsa/signed_agreements/members/1 (for the detail view)
and
cdsa/signed_agreements/members/1/update (for the update view)

Also add a form that is used in this view, tests for the form, and
tests for the view.