Rework audit such that AgreementMajorVersion is_valid is ignored
If the major version associated with a signed agreement is not valid
but the SignedAgreement is still active, still allow that
SignedAgreement to have access to the CDSA (and any workspaces associated
with that DataAffiliate CDSA's study). We decided that the view to
invalidate an AgreementMajorVersion instance would also set the
status to LAPSED for all SignedAgreements associated with that version.
This gives us additional flexibility if we need in the future (eg, we
can change a LAPSED SignedAgreement to ACTIVE if we need to extend
their access for some reason). Therefore, we only need to consider
the status of the SignedAgreement (and its associated primary, if
applicable).