Fix number_approved_dars test for change in table ordering
The test for rendering number_approved_dars for the dbGaPDataAccessSnapshot
table started failing when we changed the ordering of the table.
Change the test so is it more robust to that change by testing the
render function directly instead of pulling the value out of the
table rows and assuming an ordering. Also add a test for the number
of requested dars.