Allow DARs table to be downloaded
Use the ExportMixin in the dbGaPDataAccessRequestList view, which
lets you download the table as a text file. Add a value_dbgap_accession
method to the table class such that the dbgap_accession in the exported
table is shown without the link. Add tests to check that the link
to export the file shows up on the page and that a file is downloaded.