Subclass Column to make a dbGaP accession column
This column has custom render and value methods that display the
accession with a link to the dbGaP page, plus a bootstrap icon
indicating an external link. The column can then be reused in the
dbGaPWorkspace and dbGaPDataAccessRequest tables instead of redefining
the render_dbgap_accession method in both tables.