Handle checking groups for collaborative analysis audits
In addition to checking analyst auth domain membership, also check
that there are no groups in the auth domain. Make an exception for
PRIMED_CC_ADMINS, which we expect to be in all of the groups in the
app. TBD how to handle PRIMED member/writer groups. To facilitate
this update, change the AccessAuditResult.account attribute name to
AccessAuditResult.member and the type to be either an Account or
ManagedGroup.