Add an audit view for collaborative analysis workspaces
This view shows the results of an access audit for collaborative
analysis workspaces. Add classes to the audit source file that
enable showing this website (e.g., tables, etc.). Add a url and a
template. Note that we haven't figured out how to handle the DCC
writers group yet, so there is a test that fails with a note about
that.