Rework dbGaPDataAccessRequest tables
Instead of redefining the columns for DAR tables, instead define
them once in the original dbGaPDataAccessRequestTable, then subclass
that table for the other tables. The columns that need to change can
be redfined (or set to None) in the child tables if necessary. Also
add a link to the DAR history in some of the tables.