Use only one class for auditing dbGaP access
Instead of having an abstract base class and subclassing that to audit
a dbGaPApplication or a dbGaPWorkspace, instead just use the super
class itself. Update dbGaPAccessAudit to be a non-abstract class and
to take a queryset of dbGaPApplications and dbGaPWorkspaces to audit.
Remove the dbGaPApplicationAccessAudit and dbGaPWorkspaceAccessAudit
classes. Note that views and the rest of the app is still broken.