Add a view to handle SignedAgreement audit results
This view re-runs the audit and takes the appropriate action. In
future commits we can add htmx functionality to this view. Part of
this view meant reworking the audit structure itself. Use the same
action url for all types of AccessAuditResult and set the action
text in subclasses. Add str methods for the audit result classes
so the text can be displayed on the audit resolve page. Finally,
add a get_all_results method to the audit class that returns the
concatenation of audit results for all SignedAgreements.