Add a view to resolve CDSA workspace audits
Add a view that runs an audit on a specific CDSA workpsace and
(if a post request) handles that audit appropriately. The view can
also handle htmx requests. Modify the workspace_audit classes for
CDSA workspaces as necessary to achieve this: add a str method that
can be displayed on the detail page; set the "action" text for
different classes with a default, and change the action_url to be
the same for all the different types of audit result classes. Add a
get_all_results method to the WorkspaceAccessAudit class.