Add view to resolve a collab analysis workspace audit
Add a view that runs an audit on a specific email and
CollaborativeAnalysisWorkspace and (if a post request) handles that
audit appopriately. The view can also handle htmx requests. Modify
the workspace audit classes for CollaborativeAnalsyisWorkspaces as
necessary to achieve this: add a str method that can be displayed on
the detail page; set the "action" text for different classes with a
default, and change the action_url to be the same for all the
different types of audit result classes. Add a get_all_results method
to the CollaborativeAnalysisWorkspaceAccessAudit class.