Add additional_limitations field to DataAffiliate CDSAs
Data affiliates can specify additional limitations on what can be
done with the data outside of the workspace request form. Track that
information in the DataAffiliate model. Also add a clean method to
verify that only primary DataAffiliateAgreements can have additional
limitatoins; for now, components cannot. This could be changed in
the future.