Add view tests for DataAffiliate primary-only fields
Add tests in the DataAffiliateCreate view to verify that you can't
create a DataAffiliate component with the additional_limitations
or requires_study_review=True. These tests are failing - the clean
method isn't working because it is essentially cleaning two models
together, and the DataAffiliateAgreement form instance doesn't have
the signed_agreement field set before the form/formset are saved.