Change javascript to copy workspaces input on click
Move the button to the card subtitle instead of dynamically adding
it via the javascript (as the blog example did). Move the javascript
into the template inline_javascript block, since this code is only
relevant for this template at the moment - there's no need to put
it in the project-wide javascript. To make the javascript work,
there must be:
- an html element with the class "copy-element". This must have an
  attribute "copy-target" that specifies the id of the element whose
  text should be copied.
- an element with the id specified in the "copy-target" attribute