Move ACM env variable setting to production file
Having it in the base file was causing issues. The prod file should
detect it from the environment variable, while the test file can
just set it to a random string. Trying to detect it from the env
variable and then set it to a random string was breaking github
actions.