Allow dbGaP PIs to see DAR snapshots for their application
As part of this update, rework the new view mixin to require views
to implement a "get_dbgap_application" method. This method already
existed for dbaPDataAccessSnapshotDetail; I had to add it for the
dbGaPApplicationDetail view. Note that database queries are not
optimized here, so many of the views are making multiple queries
for the same object.