Allow dbGaP PIs to see the DAR history page
Rework the url and view such that both the dbgap_application and
the DAR id are required. This should be unnecessary (e.g., a DAR id
is only associated with a single dbGaP application at dbGaP), but it's
a little safer and it's easier to check if the user has permission
to see the page.