Allow CDSA reps, accessors, and uploaders to see detail page
Add a new viewmixin that controls access to CDSA detail pages. It
grants permission to the represenrative of the CDSA, any named
accessors or named uploaders to see the detail view for their
CDSA. Use the view mixin in the MemberAgreementDetail view, the
DataAffiliateDetail view, and the NonDataAffiliateDetail view.
Modify the information shown to users who are not CC staff by
removing links to pages they can't access and hiding some dates.