Use temporary version of ACM in requirements file
This is the version of ACM that has custom workspace adapter methods
implemented, and I'd like to test it out before releasing the new
version of the package. CI will fail because the version in the
requirements.in file does not have the version in requirements.txt.
This is good, because we need to update requirements.in with the
official released version once it's ready.