Check if authorization_domains is a field before altering it
In the WorkspaceAuthDomainDisabledForm, check if authorization_domains
is actually a field before attempting to set it to disabled. Before
this fix, attempting to set it to disabled was causing an error
in the ACM WorkspaceUpdate view, because that view removed some of
the fields that can't be changed after creation, including the
authorzations_domains field. Add tests to catch this case.