Fix failing test for mysql
test_dbgap_project_id_does_not_match was failing only in mysql likely
because in mysql tests, the pk field does not reset to 1 for each test.
A (new) previous test likely created a dbGaP application, so when this
test had the project_id hard-coded, it failed. Now, set the dbGaP
application project_id using the application that the test created
in the setup method.