Try adding a test to capture inventory inputs bug
We've found a bug associated with the inventory inputs view/helper.
From inspecting prod data, it seems to occur when there are multiple
workspaces associated with more than 1 study. I added a test that
might fix this, but it isn't failing in sqlite. I'm not sure if this
is an issue with the test or if it will only fail for mysql, so I'm
commiting and pushing to check against CI.

However I can't get it to fail on mysql