Sort workspaces before concatenating study names
The get_workspace_for_phenotype_inventory helper function had a bug
when workspaces were not properly sorted before calling itertools.groupby.
This was only apparent on the stage/prod servers, and not in mariadb.
I couldn't reproduce the bug in CI (possibly because the charset was
different between prod and stage?), and I'm not entirely sure why.
However, this fix to the code makes an interactive check in stage
work correctly.