Add phenotype inventory test to see if it fails
This new test tests a possibly related bug (see PR https://github.com/UW-GAC/primed-django/pull/675)
that we couldn't get to fail in CI, possibly due to different
collation settings. If it fails here, adding the charset/collation
setting for the test database should make it fail in this branch.