Fix CI for upload-artifact breaking change
The latest version of the upload-artifact action added a breaking
change where hidden files are not uploaded by default. Because
the coverage step was creating a hidden file (i.e., prefixed with a
dot), the file was not uploaded and the coverage job would fail
because it couldn't download the coverage file artifacts. Fix this
by:
1. making upload-artifact fail if no files were found, instead of warn
2. creating coverage files that are not hidden