Use release-drafter action draft releases and notes
Upon a commit/merge to deploy/production, this action drafts a new
release (but does not release it) and tag name (but does not tag it).
Both the release and the tag are set to the current date, or the
current date with a -1, -2 appended if a release with the current
date already exists. The release notes drafted by the action separate
out dependency updates from other changes, which will help us sort
through the release notes to see what of note has changed.