WIP - updating to use .get() instead of .first() for auth domains
This is because some workspaces now have a second auth domain that, and
we can't make the assumption that there is only one auth domain per
workspace. Using .get() ensures that the code will break if there are
two auth domains managed by the app instead of making the assumption
that there is only one and do the incorrect thing. Commiting WIP so
as to not lose work.