Add failing tests until we figured out intended behavior
We need to figure out intended behavior for the collaborative analysis
workspaces when the source workspaces have an auth domain that is
not managed by the app. Since we can't check membership, it is not
straightforward to determine who should have access. Add testst that
will fail until we figure out how it should behave, at which point
those tests can be updated.